+++ src/simple-entity/index.ts
@@ -1,5 +1,4 @@
 export {
-  SIMPLE_ENTITY_TYPES,
   SimpleEntityType,
   type SimpleEntity,
   type SimpleEntityData
@@ -8,8 +7,3 @@ export {
   convertToSimpleEntity,
   type WikiEntityToEntityOptions
 } from "./convert-to-simple-entity";
-export { getEntityCountryCode } from "./get-entity-country";
-export { getEntityData } from "./get-entity-data";
-export { getEntityTypeByExtract } from "./get-entity-type-by-extract";
-export { getEntityType as getEntityTypeByOntology } from "./get-entity-type";
-export { getEntityType as getEntityInstanceType } from "./get-entity-instance-type";